# [SYCL] Improve the next-token speed of Q4_0