- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [ ] Low
  - [x] Medium
  - [ ] High

## Changes:
1. Add a cache-friendly layout of Q4_0 weight.
2. Write a new dmmv kernel for Q4_0 weight.
3. Support dmmv for 1<row
4. Use dmmv as default for 9>prompt_len>0, it's faster than mmvq.
5. 

## Performance:
llama-2-7b.Q4_0.gguf on 12700H+A770m
```
master branch:
 Once you have decided to start a business, you will need to decide what type of business you want to start. Unterscheidung zwischen einem Unternehmen und einer Person.
The
llama_print_timings:        load time =    3462.20 ms
llama_print_timings:      sample time =       0.37 ms /    32 runs   (    0.01 ms per token, 86720.87 tokens per second)
llama_print_timings: prompt eval time =      88.22 ms /     2 tokens (   44.11 ms per token,    22.67 tokens per second)
llama_print_timings:        eval time =     777.04 ms /    31 runs   (   25.07 ms per token,    39.90 tokens per second)
llama_print_timings:       total time =     866.21 ms /    33 tokens
Log end

```
```
PR branch:
 Once you have decided to start a business, you will need to decide what type of business you want to start. Unterscheidung zwischen einem Unternehmen und einer Person.
The
llama_print_timings:        load time =    5027.14 ms
llama_print_timings:      sample time =       0.38 ms /    32 runs   (    0.01 ms per token, 83989.50 tokens per second)
llama_print_timings: prompt eval time =      36.53 ms /     2 tokens (   18.26 ms per token,    54.75 tokens per second)
llama_print_timings:        eval time =     578.50 ms /    31 runs   (   18.66 ms per token,    53.59 tokens per second)
llama_print_timings:       total time =     616.66 ms /    33 tokens
Log end

```
eval performance from 39.90 tokens/s to 53.59 tokens/s, ~34% end2end speedup.
prompt(len<=8) eval performance from 22.67 tokens/s to 54.75 tokens/s, ~140% end2end speedup.